requirements.txt
flask==3.1.1
flask-cors==6.0.0
opencv-python-headless==4.11.0.86
numpy==2.2.6
pillow==11.2.1
dlib==20.0.0
waitress==3.0.2
tensorflow==2.19.0
scikit-learn==1.6.1
joblib==1.5.1
matplotlib==3.10.3
firebase-admin==6.2.0
python-dotenv==1.0.0
psutil==5.9.8
requests==2.31.0